requirements.txt
Flask>=2.3,<4.0
flask-cas-ng>=1.0.1
gunicorn>=21.2
